deps/v8/test/inspector/debugger/framework-precise-ranges-expected.txt

Checks framework debugging with blackboxed ranges.

Running test: testEntireScript
{
    id : <messageId>
    result : {
    }
}

Running test: testFooNotBlackboxed
{
    id : <messageId>
    result : {
    }
}
foo (test.js:8:12)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

foo (test.js:9:2)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

foo (test.js:9:15)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)


Running test: testFooBlackboxed
{
    id : <messageId>
    result : {
    }
}
testFunction (test.js:14:21)
(anonymous) (expr.js:0:0)

testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

boo (test.js:12:2)
foo (test.js:9:9)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

boo (test.js:12:12)
foo (test.js:9:9)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

testFunction (test.js:16:0)
(anonymous) (expr.js:0:0)


Running test: testBooPartiallyBlackboxed1
{
    id : <messageId>
    result : {
    }
}
foo (test.js:8:12)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

foo (test.js:9:2)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

boo (test.js:12:2)
foo (test.js:9:9)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

boo (test.js:12:12)
foo (test.js:9:9)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

foo (test.js:9:15)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)


Running test: testBooPartiallyBlackboxed2
{
    id : <messageId>
    result : {
    }
}
foo (test.js:8:12)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

foo (test.js:9:2)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

boo (test.js:12:2)
foo (test.js:9:9)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

boo (test.js:12:12)
foo (test.js:9:9)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

foo (test.js:9:15)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)


Running test: testBooPartiallyBlackboxed3
{
    id : <messageId>
    result : {
    }
}
foo (test.js:8:12)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

foo (test.js:9:2)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

boo (test.js:12:2)
foo (test.js:9:9)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

boo (test.js:12:12)
foo (test.js:9:9)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)

foo (test.js:9:15)
testFunction (test.js:15:2)
(anonymous) (expr.js:0:0)
